Moving tunnel.py code into protocol.proxy for better separations of concerns.
