+++ tests/tunnel_tests.py
@@ -89,28 +89,63 @@ class ProxyBufferTest(testing.AsyncTestCase):
         buffer = proxy.ProxyBuffer(5)
         bytes_written = await buffer.enqueue(b'abcd')
         self.assertEqual(4, bytes_written)
-        print(buffer.contents)
+        total = await buffer.get_total_bytes_enqueued()
+        self.assertEqual(4, total)
+        total = await buffer.get_total_bytes_dequeued()
+        self.assertEqual(0, total)
+        size = await buffer.get_current_count()
+        self.assertEqual(4, size)
+
         data = bytearray(2)
         bytes_read = await buffer.dequeue(data)
         self.assertEqual(2, bytes_read)
         self.assertEqual(b'ab', bytes(data))
+        total = await buffer.get_total_bytes_enqueued()
+        self.assertEqual(4, total)
+        total = await buffer.get_total_bytes_dequeued()
+        self.assertEqual(2, total)
+        size = await buffer.get_current_count()
+        self.assertEqual(2, size)
 
         bytes_read = await buffer.dequeue(data)
         self.assertEqual(2, bytes_read)
         self.assertEqual(b'cd', bytes(data))
+        total = await buffer.get_total_bytes_enqueued()
+        self.assertEqual(4, total)
+        total = await buffer.get_total_bytes_dequeued()
+        self.assertEqual(4, total)
+        size = await buffer.get_current_count()
+        self.assertEqual(0, size)
 
         bytes_written = await buffer.enqueue(b'efg')
         self.assertEqual(3, bytes_written)
+        total = await buffer.get_total_bytes_enqueued()
+        self.assertEqual(7, total)
+        total = await buffer.get_total_bytes_dequeued()
+        self.assertEqual(4, total)
+        size = await buffer.get_current_count()
+        self.assertEqual(3, size)
 
         bytes_read = await buffer.dequeue(data)
         self.assertEqual(2, bytes_read)
         self.assertEqual(b'ef', data)
+        total = await buffer.get_total_bytes_enqueued()
+        self.assertEqual(7, total)
+        total = await buffer.get_total_bytes_dequeued()
+        self.assertEqual(6, total)
+        size = await buffer.get_current_count()
+        self.assertEqual(1, size)
+
         bytes_read = await buffer.dequeue(data)
         self.assertEqual(1, bytes_read)
         self.assertEqual(b'g', data[0:1])
+        total = await buffer.get_total_bytes_enqueued()
+        self.assertEqual(7, total)
+        total = await buffer.get_total_bytes_dequeued()
+        self.assertEqual(7, total)
+        size = await buffer.get_current_count()
+        self.assertEqual(0, size)
 
-
-        print("DEQUEUED: {}".format(data))
 
 @unittest.skip('Not ready')
 class WebsocketServerTest(testing_utils.AsyncWsproxyTestCase):
