+++ wsproxy/routes/registry.py
@@ -3,8 +3,9 @@
 Registry of the (JSON) routes in the system.
 """
 import itertools
+from wsproxy.protocol.proxy import get_routes as get_proxy_routes
 from wsproxy.routes import (
-    info, socks5, tunnel
+    info, socks5
 )
 
 
@@ -13,6 +14,6 @@ def get_route_mapping():
     return {
         route.name: route
         for route in itertools.chain(
-            info.get_routes(), tunnel.get_routes(), socks5.get_routes()
+            get_proxy_routes(), info.get_routes(), socks5.get_routes()
         )
     }
